% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialGO.R
\name{buildEdgeDF}
\alias{buildEdgeDF}
\title{Build a data frame of edges from the nodes. NOTE: Due to a bug
in the way DiagrammeR renders Graphviz objects, edges currently do not render
with the correct stylings. To see that they are specified correctly, one can
render the graph with the output="visNetwork" option, which shows correct
edges but loses the node layout}
\usage{
buildEdgeDF(nodeDF)
}
\arguments{
\item{nodeDF}{A data frame containing at least columns id, label,}
}
\value{
A data frame of edges (from, to, rel)
}
\description{
Build a data frame of edges from the nodes. NOTE: Due to a bug
in the way DiagrammeR renders Graphviz objects, edges currently do not render
with the correct stylings. To see that they are specified correctly, one can
render the graph with the output="visNetwork" option, which shows correct
edges but loses the node layout
}
\examples{
\dontrun{
scores <- c(0.5, 0.2, 0.001)
names(scores) <- c("GO:0008150", "GO:1901360", "GO:0006139")
buildEdgeDF(buildNodeDF(scores, 2))
}
}
