% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialGO.R
\name{buildLabel}
\alias{buildLabel}
\title{Create a text label describing the GO term and its score}
\usage{
buildLabel(goTerm, scores)
}
\arguments{
\item{goTerm}{A string containing the GO ID to use}

\item{scores}{A named numerical vector containing GO IDs as names
and the corresponding score}
}
\value{
A string containing the label text
}
\description{
Create a text label describing the GO term and its score
}
\examples{
\dontrun{
scores <- c(0.5, 0.2, 0.001)
names(scores) <- c("GO:0008150", "GO:1901360", "GO:0006139")
buildLabel(names(scores[1]), scores)
}
}
