% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialGO.R
\name{buildNodeDF}
\alias{buildNodeDF}
\title{Builds a node data frame from the top GO terms using the enrichment results}
\usage{
buildNodeDF(enrichment_results, top)
}
\arguments{
\item{enrichment_results}{A named numerical vector containing GO IDs as names
and the corresponding score}

\item{top}{The number of top nodes to use to build the graph}
}
\description{
Builds a node data frame from the top GO terms using the enrichment results
}
