% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{checkInput}
\alias{checkInput}
\title{Check input values for correct format and type}
\usage{
checkInput(enrichment_results, top, cutoff)
}
\arguments{
\item{enrichment_results}{A data frame to be used for graph generation}

\item{top}{The number of nodes to use when generating the subgraph}

\item{cutoff}{The cutoff to be used for hiding nodes}
}
\value{
TRUE if the input is correct, FALSE otherwise
}
\description{
Check input values for correct format and type
}
