% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{generateGraph}
\alias{generateGraph}
\title{Generate the graph visualization of GO enrichment scores}
\usage{
generateGraph(scores, top, cutoff)
}
\arguments{
\item{scores}{A named numeric vector with GO IDs and corresponding scores}

\item{top}{The number of nodes to use when generating the subgraph}

\item{cutoff}{The cutoff to be used for hiding nodes}
}
\value{
A graph object to be used with render_graph()
}
\description{
Generate the graph visualization of GO enrichment scores
}
\examples{
\dontrun{
scores <- c(0.5, 0.2, 0.001)
names(scores) <- c("GO:0008150", "GO:1901360", "GO:0006139")
generateGraph(scores, 2, 0.3)
}

}
