% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{getEdgeColor}
\alias{getEdgeColor}
\title{Determine what color the edge should take in the graph.}
\usage{
getEdgeColor(edgeRelationships)
}
\arguments{
\item{edgeRelationships}{A string containing a relationship between GO terms}
}
\value{
A string containing the color the edge should take
}
\description{
Determine what color the edge should take in the graph.
}
\examples{
\dontrun{
getEdgeColor("is_a")
}

}
