% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{getSubgraphNodes}
\alias{getSubgraphNodes}
\title{Generate the subgraph implied by the list of nodes passed in.}
\usage{
getSubgraphNodes(enrichment_results)
}
\arguments{
\item{enrichment_results}{A named numerical vector containing GO IDs as names
and the corresponding score}
}
\value{
A data frame containing one column, named "id", with the nodes needed
        for building the graph
}
\description{
Generate the subgraph implied by the list of nodes passed in.
}
\examples{
\dontrun{
scores <- c(0.5, 0.2, 0.001)
names(scores) <- c("GO:0008150", "GO:1901360", "GO:0006139")
getSubgraphNodes(scores)
}

}
