% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go_ids}
\alias{go_ids}
\title{ALL GO enrichment data}
\format{A named numerical vector containing GO IDs as names and their score
in the enrichment analysis as values.}
\usage{
data(go_ids)
}
\description{
Data from a GO enrichment analysis of dataset of Acute Lymphoblastic Leukemia
using topGO. This dataset was retrieved from the "ALL" library.
}
\examples{
data(go_ids)
\dontrun{
render_graph(generateGraph(go_ids, 5, 0.2))
}
}
\references{
Chiaretti, S. et al. Gene expression profile of adult T-cell
acute lymphocytic leukemia identifies distinct subsets of patients with
different response to therapy and survival. Blood 103, 2771–2778 (2004).
}
\keyword{datasets}
