% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{rebuildGOID}
\alias{rebuildGOID}
\title{Rebuild a GO ID string from its number}
\usage{
rebuildGOID(goID)
}
\arguments{
\item{goID}{A string or number containing the GO ID number}
}
\value{
The full GO ID
}
\description{
Rebuild a GO ID string from its number
}
\examples{
rebuildGOID(8150)

}
