% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialGO.R
\name{reduceGraph}
\alias{reduceGraph}
\title{Removes nodes from the graph which have a p-value above the specified cutoff,
collapsing the edges}
\usage{
reduceGraph(graph, cutoff)
}
\arguments{
\item{graph}{The graph to reduce}

\item{cutoff}{The cutoff to be used (nodes above this cutoff will be removed)}
}
\value{
A graph with the nodes removed
}
\description{
Removes nodes from the graph which have a p-value above the specified cutoff,
collapsing the edges
}
