% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{server}
\alias{server}
\title{Shiny app server function}
\usage{
server(input, output)
}
\arguments{
\item{input}{The input object}

\item{output}{The output object}
}
\description{
Shiny app server function
}
