% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialgo.R
\name{val2col}
\alias{val2col}
\title{This function converts a vector of values("z") to a vector of color
levels. One must define the number of colors. The limits of the color
scale("zlim") or the break points for the color changes("breaks") can
also be defined. when breaks and zlim are defined, breaks overrides zlim.}
\usage{
val2col(z, zlim, col = heat.colors(12), breaks)
}
\description{
Function copied from https://menugget.blogspot.com/2011/09/converting-values-to-color-levels.html
}
